requirements.txt
altgraph==0.17.3
attrdict3==2.0.2
attrs==22.2.0
boardgamegeek2 @ git+https://github.com/Conday/boardgamegeek.git@f422467fac65a12fcf720a206e7f0ce7c12cd20a
cairocffi==1.4.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colored==1.4.4
cssselect2==0.7.0
defusedxml==0.7.1
exceptiongroup==1.1.0
idna==3.4
lxml==4.9.2
numpy==1.24.2
Pillow==9.4.0
platformdirs==3.1.0
psutil==5.9.4
pycparser==2.21
pygtrie==2.5.0
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.0
pypdf==3.5.1
reportlab==3.6.12
requests==2.28.2
requests-cache==1.0.0
six==1.16.0
svg-stack==0.1.0
svglib==1.5.1
svgutils==0.3.4
tinycss2==1.2.1
tqdm==4.65.0
url-normalize==1.4.3
urllib3==1.26.14
webencodings==0.5.1
